seperate each bindings to js file
